requirements.txt
numpy==1.16.3
Pillow==6.0.0
